requirements.txt
asgiref~=3.2.10
bokeh~=2.1.1
certifi~=2019.9.11
chardet==3.0.4
click~=7.0
decouple==0.0.7
dj-database-url==0.5.0
Django~=3.0.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-extensions==2.2.9
django-heroku==0.3.1
django-rest-framework==0.1.0
gunicorn==20.0.4
idna~=2.8
Jinja2~=2.10.1
MarkupSafe==1.1.1
numpy~=1.17.2
packaging==20.4
pandas~=0.25.1
Pillow~=7.2.0
psycopg2==2.8.5
pycoingecko~=0.4.0
pyparsing==2.4.7
python-dateutil~=2.8.0
pytz~=2019.2
PyYAML==5.3.1
redis==3.5.3
requests~=2.22.0
rq==1.4.2
six~=1.10.0
sqlparse~=0.3.0
tornado~=6.0.3
typing-extensions==3.7.4.2
Unipath==1.1
urllib3~=1.25.6
whitenoise==5.1.0
djangorestframework~=3.11.0
cffi~=1.12.3
pip~=20.1.1
request~=2019.4.13
cryptography~=2.9.2
lxml~=4.4.1
notebook~=6.0.1
colorama~=0.4.1
setuptools~=41.2.0
Twisted~=19.7.0
Pygments~=2.4.2
channels~=2.4.0
django-debug-toolbar~=2.2